+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [2.1.1] - 2024-10-06
+### Added
+- Added on_status hook to `execute!`
 ### Changed
 - Reuse same attribute instance